requirement.txt
flask==2.2.5
opencv-python==4.7.0.72
deepface==0.0.83
numpy<2
tensorflow==2.10.0
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
bcrypt==4.0.1
pymysql==1.1.0
cryptography==41.0.7
redis==5.0.1
schedule==1.2.0
Pillow==10.0.1